structs1.rs with String ala Rust book
Currently, the Rust Book doesn't introduce lifetimes until Chapter 10,
so the current solution with &'a str is quite foreign for students
following it.

So, for people who are using the Book, here is an example that follows it.

I just noticed that the structs README actually points to the
chapter on structs from Rust by Example

 https://doc.rust-lang.org/rust-by-example/custom_types/structs.html

which notes the use of lifetimes without really discussing them, just providing
a link to the chapter on them.